Google Books API & Profile lists

+++ database/migrations/2023_03_15_112145_create_books_table.php
@@ -25,7 +25,7 @@ class CreateBooksTable extends Migration
             $table->string('img',1000)->nullable();
             $table->string('date')->nullable();
             $table->integer('pages')->nullable();
-            $table->integer('rate')->nullable();
+            $table->float('rate')->nullable();
             $table->boolean('readBefore')->nullable()->default(false);
             $table->boolean('searchable')->default(true);
             $table->string('link')->nullable();
